transaction : board update

+++ SpringBoard/src/main/java/com/ict/mapper/BoardMapper.java
@@ -34,6 +34,6 @@ public interface BoardMapper {
 	
 	// 댓글이 써질때마다 board_tbl의 댓글 개수 업데이트하기
 	public void updateReplyCount(@Param("bno") Long bno,
-									@Param("amount") Long amount);
+									@Param("amount") int amount);
 	
 }
